Improve upload list (#4516)
* support prevent removing when return value is false, close #4456

* Add showRemoveIcon & showPreviewIcon into showUploadList

close #4384 & #4351

* fix typo

+++ components/upload/index.tsx
@@ -189,16 +189,17 @@ export default class Upload extends React.Component<UploadProps, any> {
   }
 
   handleRemove(file) {
-    const onRemove = this.props.onRemove;
-    if (onRemove) {
-      onRemove(file);
-    }
-    let fileList = this.removeFile(file);
-    if (fileList) {
-      this.onChange({
-        file,
-        fileList,
-      });
+    const { onRemove } = this.props;
+    // Prevent removing file
+    const onRemoveReturnValue = onRemove && onRemove(file);
+    if (onRemoveReturnValue !== false) {
+      let fileList = this.removeFile(file);
+      if (fileList) {
+        this.onChange({
+          file,
+          fileList,
+        });
+      }
     }
   }
 
@@ -251,12 +252,15 @@ export default class Upload extends React.Component<UploadProps, any> {
     });
     delete rcUploadProps.className;
 
+    const { showRemoveIcon, showPreviewIcon } = showUploadList as any;
     const uploadList = showUploadList ? (
       <UploadList
         listType={listType}
         items={this.state.fileList}
         onPreview={onPreview}
         onRemove={this.handleManualRemove}
+        showRemoveIcon={showRemoveIcon}
+        showPreviewIcon={showPreviewIcon}
       />
     ) : null;
